Do not reset Patch History selection on navigation to next file diff
When the user views a file diff and then changes in the Patch History
the selection of side A (left side), then this selection is lost when
the user navigates to the previous or next file diff. The selection of
side A is always reset to the patch set that is selected as 'Old
Version History' on the ChangeScreen (the base for comparing patch
sets).

This change ensures that the selection in the Patch History is kept
when the user navigates to the previous or next file diff.

Bug: issue 999
Change-Id: I19cafedd8a11683dca4a24fbebbcacf63425fe62
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: c82e9e98b6d41dbd2267d72b96627a03c94739ab