Rename "-- All Projects --" to "All-Projects"
The name "-- All Projects --" is awkard to work with on the command
line due to starting with leading hypens, and containing spaces.

In prior versions of Gerrit Code Review this did not really matter,
the project was stored only in the database and did not need to be
accessed by command line tools. Now that the access controls are
in Git and are stored in a Git repository, this directory exists
and needs to be managed.

Bug: issue 994
Change-Id: I19b0967c6cf2720dcd2fd72a354e3566def15fdb
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 99071dd23240d121967f84759651f61f698442e0