Replace the shell code that tries to emulate git config with an error.
Replace the code with an error message that git has to be
in the PATH instead of trying to parse gerrit.conf with
shell code. The shell code do not remove whitespaces as
it should and this will give stranger errors when looking
for java later in the script (all platforms/shells).
This code doesn't seem to have been used anyway.

Change-Id: Ideba5b950a97c0e77839cbddd4616ef737470da2

Former-commit-id: f86b2141843e3dced579330dd8e11c5ea01f9dd4