Fix NPE in PermissionRuleEditor when group lacks UUID
If a group does not have an entry in the "groups" table within
the refs/meta/config branch its UUID will be null, and a crash
used to occur in the UI trying to build the group name as a link.

Instead render the group name as a span, without the link.

Bug: issue 1002
Change-Id: Ia43f1950f53427acbc024543508c5e4c248f73bf

Former-commit-id: e7fdc9aeda12402fa9abce89ea6f4d046a60235a