create-project: Bring back --permissions-only flag
If a project is permissions only, assign HEAD to point to
refs/meta/config. This way gitweb view of the project shows the
permissions history by default, and clients that clone the project
are able to get a detached HEAD pointing to the current permission
state, rather than an empty repository.

Change-Id: I194bfef1af9d00268de9c503ea986879f1fbaff9

Former-commit-id: aaca9593f13cca311f5041e61317ec190353a854