ls-projects: Refactor --tree support to shorten code
I found the old way this was organized harder to read. Since I am
about to add a bunch of new logic to filter around permissions only
projects, lets move the showTree related stuff earlier so its out
of the way before we get to the filtering and branch output.

Change-Id: I2f4fc6a0fb1d9919048649c89bcc7e60a3bc64bc

Former-commit-id: 8979aa21fc2260822c23c366c20a077a99e90da4