Fix MySQL counter resets
MySQL resets its counters to 0 on server restarts, resulting in
database errors on the next insertion attempt for common entities
like account or change. gwtorm 1.1.5 was patched to leave in the
dummy row that incremented the counter, ensuring the server will
use MAX() + 1 instead of 1 on the next increment after restart.

Bug: issue 518
Change-Id: I52cf5354f4fccf787fe8c62539ead6d5f6a0fead

Former-commit-id: 17ca55ff0f4553212c9fc53275d7cf75a0b6d188