Make Gerrit with gwt 2.3.0 run in gwtdebug mode
Running Gerrit in gwtdebug mode resulted on client side in an
exception saying that javax.validation.ContraintViolation could not be
resolved. As a result of this exception the project access screen
failed to load.

This change now adds Maven dependencies to the validation-api jars in
the pom.xml of the gerrit-gwtdebug project to this fix this problem.

Change-Id: I8b2341c017a6cfbde490bd38b4e016d29707f96c
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: e2900ea600f4ffe12dfae99612b8f46622b7b4fe