Update com.jcraft.jsch to 0.1.44-1
This is the version also used by JGit since 74419e14. According to [1]
the version 0.1.44 was empty when uploaded to maven central, as maven
artifacts uploaded to maven central are immutable a new version 0.1.44-1
was created to fix this. It seems local builds found other maven
repositories containing a corrected 0.1.44 version. As this version has
been removed from maven central we should use 0.1.44-1 instead. This
should also fix the gerrit CI build running at [2].

[1] https://issues.sonatype.org/browse/OSSRH-1099
[2] http://ci.jenkins-ci.org/job/gerrit_master/

Change-Id: I008ae0e7d3f13844a89330297f81fa057f47134d
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>

Former-commit-id: 32ace8139b217f7d34a1886b866004b58564e045