Fix non-volatile fields in ProjectState
Both of these fields are supposed to be volatile, but for unknown
reasons I marked them transient. They both need to be volatile as
the fields are accessed by concurrent threads. The use sites are
already prepared for these to be volatile, they just were not set
correctly in the declaration.

Change-Id: Icffb185487e1bf8f75358257a2d68e21ae90b08f

Former-commit-id: f013fb11525104493db8a4fa0af61ddc09d11996