Add menu to AccountGroupScreen
The AccountGroupScreen provides many functions, renaming the group,
setting the group owner, setting the group description, setting the
group options, setting the group type and setting the group members
and included groups. As a result the screen seems to be overloaded
with controls. To improve the overview this change introduces a menu
in the AccountGroupScreen and different screens for subsets of the
functionality (similar as it's done for the ProjectScreen). For now
there are two screens:
1. AccountGroupInfoScreen: provides renaming the group, setting the
group owner, setting the description, setting the group options and
the group type
2. AccountMembersScreen: provides setting the group members and
included groups

Links from other screens to the AccountGroupScreen are resolved
depending on the group type. For Gerrit internal groups we show as
default the AccountGroupMembersScreen since most users will be
interested in the group members when they navigate to the group. For
external and system groups Gerrit cannot display the group members,
this is why for these groups by default the AccountGroupInfoScreen is
shown.

Change-Id: I19e4df01224510bf6af18e5cd6675381df07f13d
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: dbd05783162ad9798010063b58726635b8450009