Use remote.name.replicatePermissions to hide permissions
Administrators can now disable replication of permissions-only
projects and the per-project refs/meta/config in replication.config:

  [remote "backup"]
    url = backup-server:git/${name}.git
    replicatePermissions = true

  [remote "public"]
    url = public-mirror:/pub/git/${name}.git
    authGroup = Public Mirror
    replicatePermissions = false

By default replicatePermissions is true, as most of the permission
data is public by default in the web UI and the replication feature
was primarily built to backup a server to its warm-spares or slaves.
If set to false, permissions only projects and the refs/meta/config
branch of each individual project will not be replicated.

Bug: issue 813
Change-Id: I229fdca08277c1b0c61dc10c3591a5f8f1bc9f35

Former-commit-id: 0028d114e89e7ab2ee50c9a22f321a989223dc42