Add additional information in Groups Table.
Add columns for 'Owners', 'Group Type', 'Email Only Authors' and
'Visible To All' into the groups table used by Admin->Groups Screen
and MyGroups Screen. Adding more info to this table helps the
administrators to view the settings and additional info of a group
faster without clicking on each group. This can work as a dashboard
or report that contains the info for all the groups in one place.
As an example, this allows one to quickly find all the groups which
are having emails squelched.

Change-Id: Ia952e9dcecf415c4fce7001a0d28ad4731ff5b1c

Former-commit-id: fccf928042d372f379204fcef61038f94c5b0845