Add a "Save" button to the PatchScriptSettingsPanel
The "Update" button now only updates the display. For
non-logged in users it behaves as it used to.  However,
for logged in users, the "Update" button no longer saves
the user preferences, and all of its settings are reset
to the saved values whenenver the PatchScreen is reloaded
after the update.  Additionaly, for logged in users, a
"Save" button now appears to the right of the "Update"
button and now behaves the way that "Update" used to
behave for logged in users.

Change-Id: I6ee5d688fbcee268aef1de1408762904d2a3c1ea

Former-commit-id: 61838dcc1ce7aa72b412c4e182a8fdc8b596dc24