Fixed a bug in PermissionRule class which broke the Schema_53.
After the commit 76409cf7 the PermissionRule.setDeny(boolean newDeny)
method was broken as the newDeny argument wasn't honored any more.
The Schema_53 made use of this method and was broken after this change.

This change removes the parameter from the setDeny method which makes its
meaning clear. This is also necessary as in the near future we will have a
new Action value "BLOCK" and then setDeny(false) would become really
unintuitive.

Change-Id: I6fd685b13c5854461a19e19bc5b74c0fc256db8a

Former-commit-id: ac3bad068086c6ab0ca644e3a9277769172454c6