Renaming getDeny() to isDeny() in the PermissionRule class.
Since getDeny returns boolean value it is more readable when named
isDeny. In a follow up change a new method "isBlock" will be
introduced and this change is just preparing this class for this new
method.

Unfortunately, renaming the getForce, another boolean method, to
hasForce, breaks the build of the gerrit-gwtui component.  The reason
must be that the property 'force' from the PermissionRuleEditor.ui.xml
gets resolved by using the "getForce" method. Leaving this method name
unchanged.

Change-Id: I1f7d99ed8148dc336e3e35369f04169c905bc0ba

Former-commit-id: e5f0457942ef2c09fe3aa7670326213350c4f861