Add inheritance of prolog rules
Projects now inherit the prolog rules defined in their parent project.
Submit results from the child project are filtered by the parent
project using the filter predicate defined the parent's rules.pl.
The results of the filtering are then passed up to the parent's parent
and filtered. This process repeats up to the top level All-Projects.
If a parent does not have a filter rule defined, the results are not
filtered.

Change-Id: I961d10149f62996d1c3c74c80935b2f72ea0acc7

Former-commit-id: 82c088e282e5e433cb4eb7dc9af0dfbac82e2162